atd-toolbox/isolate_vz_data_differences/requirements.txt
psycopg2-binary==2.9.9
tqdm==4.66.4
visidata==3.0.2
